requirements.txt
openai>=1.0.0
pydantic>=2.0.0
tenacity>=8.0.0
docker>=7.0.0
python-dotenv>=1.0.0
Flask>=2.0.0
